InformationCollectionStruct moved to Payload directory
