+++ Form/Type/InformationCollectionType.php
@@ -4,7 +4,7 @@ namespace Netgen\Bundle\EzFormsBundle\Form\Type;
 
 use Symfony\Component\Form\FormBuilderInterface;
 use Netgen\Bundle\EzFormsBundle\Form\DataWrapper;
-use Netgen\Bundle\EzFormsBundle\API\Repository\Values\InformationCollection\InformationCollectionStruct;
+use Netgen\Bundle\EzFormsBundle\Form\Payload\InformationCollectionStruct;
 use eZ\Publish\API\Repository\Values\ContentType\ContentType;
 use Netgen\Bundle\EzFormsBundle\Form\FieldTypeHandlerRegistry;
 use Symfony\Component\Form\DataMapperInterface;
@@ -62,7 +62,7 @@ class InformationCollectionType extends AbstractContentType
         if ( !$payload instanceof InformationCollectionStruct )
         {
             throw new RuntimeException(
-                "Data payload must be an instance of Netgen\\Bundle\\EzFormsBundle\\API\\Repository\\Values\\InformationCollection\\InformationCollectionStruct"
+                "Data payload must be an instance of Netgen\\Bundle\\EzFormsBundle\\Form\\Payload\\InformationCollectionStruct"
             );
         }
